[evo] xioh: Load e1000.ko
This is kinda dirty, but I can't get my head around how initramfs are
supposed to be generated for debian-installer, so I'll just drop our
updated e1000 at the root and try to load it in the early script; this
should get the install working on XiVO IPBX Open Hardware. It should
also be OK on other machines, because our e1000 should support the other
e1000-supported network cards as well.