apt-get dist-upgrade before rebooting
reason: currently, the installer installs linux-image-4.9.0-4 (from
stretch distribution), but after rebooting, apt-get dist-upgrade
installs linux-image-4.9.0-5 (from stretch-security distribution). Then
the postinst installs dahdi-linux-modules-4.9.0-4, which is correct for
the first reboot, but will fail after the next reboot, where 4.9.0-5
will be loaded but will not have the correct dahdi-linux-module.

To fix this, we apt-get dist-upgrade before rebooting, this way the
linux-image will be correct in the first place.